airflow_pipelines/requirements.txt
# Apache Airflow and providers
apache-airflow>=2.7.0
apache-airflow-providers-google>=10.0.0
apache-airflow-providers-postgres>=5.0.0
apache-airflow-providers-ssh>=3.0.0
apache-airflow-providers-slack>=8.0.0
apache-airflow-providers-openlineage>=1.0.0

# Google Cloud
google-cloud-bigquery>=3.0.0

# ClickHouse
clickhouse-driver>=0.2.0
clickhouse-connect>=0.6.0

# Data processing
pandas>=2.0.0
python-dateutil>=2.8.0
requests>=2.28.0

# GDPR anonymization
Faker>=18.0.0

# Development & Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.0.0

# Linting
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
